新增ScenePoll

+++ H6Game/H6Game.Base/Base/ComponentIdCreator.cs
@@ -4,12 +4,12 @@ namespace H6Game.Base
 {
     public class ComponentIdCreator
     {
-        private static long id = 0;
-        public static uint CreateId()
+        private static int id = 0;
+        public static int CreateId()
         {
             Interlocked.Increment(ref id);
-            Interlocked.CompareExchange(ref id, 0, uint.MaxValue);
-            return (uint)id;
+            Interlocked.CompareExchange(ref id, 0, int.MaxValue);
+            return id;
         }
     }
 }
